move dataConnection up to Code component
