+++ client/src/interface.ts
@@ -1,3 +1,5 @@
+import { DataConnection } from "peerjs";
+
 export interface HomeProps{
     setUsername: React.Dispatch<React.SetStateAction<string | null>>
 }
@@ -16,10 +18,10 @@ export interface ChangeObject {
 
 export interface CodeEditorProps {
     peerId: string
-    setChanges: React.Dispatch<React.SetStateAction<ChangeObject[]>>
     editorRef: React.MutableRefObject<any>
     codeRef: React.MutableRefObject<any>
     setCodeTrigger: React.Dispatch<React.SetStateAction<boolean>>
+    dataConn: DataConnection | null
 }
 
 export interface CodeData {
@@ -36,10 +38,5 @@ export interface ChangeObject {
 
 export interface PeerChatProps {
     peerId: string,
-    setPeerId: React.Dispatch<React.SetStateAction<string>>
-    changes: ChangeObject[],
-    setChanges: React.Dispatch<React.SetStateAction<ChangeObject[]>>
-    editorRef: React.MutableRefObject<any>
-    codeRef:React.MutableRefObject<any>
     username: string | null
 }
